api/pyproject.toml
[project]
name = "numbat-tasks-api"
version = "0.16.1.dev0"
description = "Backend API for Numbat Tasks"
author = "Michael Van Delft <michael@xo.tc>"
license = "MIT"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-nested-routers>=0.95.0",
    "drf-spectacular[sidecar]>=0.29.0",
    "psycopg[binary]>=3.3.2",
    "python-decouple>=3.8",
]

[dependency-groups]
dev = [
    "ipython>=9.9.0",
    "pytest-django>=4.11.1",
    "ruff>=0.14.11",
]
production = [
    "gunicorn>=23.0.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "numbat_tasks_api.settings"

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # flake8-bandit
    "S",
    # isort
    "I",
    # pep8-naming
    "N",
    # datetimez (ban the usage of unsafe naive datetime)
    "DTZ",
]
# Ignore line lenght as that is handled else where.
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
# Ignore "No assertions" rule in tests.
"test*.py" = ["S101", "S311"]
